refactor(v010): carve voicegw mcp out of _legacy (final command carve-out)
Thirteenth and final per-command carve-out in section 2.

Adds voicegateway/cli/mcp.py owning the 'mcp' command. Body unchanged
from v0.0.5; the 'from voicegateway.mcp.server import serve_http,
serve_stdio' import stays inside the function body so the cli does
not eagerly load the optional MCP extras for unrelated subcommands.

Every v0.0.5 command now lives in its own focused submodule. _legacy.py
is down to 71 LOC of pure infrastructure (app/console singletons, the
--version callback, _load_gateway, _parse_iso_date_arg). The next
section-2 task moves those into __init__.py and a new _helpers.py
module, then deletes _legacy.py outright.

13 commands still register on the shared app; pytest tests/test_cli.py
46/46 pass; ruff and mypy clean across the cli package (14 source
files now).
