+++ voicegateway/cli/__init__.py
@@ -30,6 +30,7 @@ from voicegateway.cli import dashboard as _dashboard  # noqa: F401, E402
 from voicegateway.cli import export_costs as _export_costs  # noqa: F401, E402
 from voicegateway.cli import init as _init  # noqa: F401, E402
 from voicegateway.cli import logs as _logs  # noqa: F401, E402
+from voicegateway.cli import mcp as _mcp  # noqa: F401, E402
 from voicegateway.cli import projects as _projects  # noqa: F401, E402
 from voicegateway.cli import reconcile as _reconcile  # noqa: F401, E402
 from voicegateway.cli import rotate_secret as _rotate_secret  # noqa: F401, E402
